Massive rewrite to account for Derpibooru updating how searches are done.
A lot of the options that previously were included as html parameters are now utilized through the search query itself. Only search format and search direction remain as image specific parameters.

This means that the program has a lot less options, but requires the user to create the new search queries themselves. Links have been provided to the default search page, and search help page to give a list of possible queries that can be used.

Various other little changes, like the version in the window title now showing up, and informing the user that custom filters will not work when an api key is provided.