Settings are now stored in a local .ini file instead of the default system location.
Previously, this caused issues with portability, as settings were stored on the registry on windows.
The program will first check for settings in the local .ini file. If no settings are found, it will check the old default location (the registry, on windows). If there are no settings there either, default settings are used. Settings are always saved to the new location, no matter where they were before.